update(plugin): introduce TablesInput
Instead of using a raw ss_plugin_init_input as the input to a plugin,
introduce a wrapper struct that does some initial validation, so we
don't need to check the `Option`s all the time.

Also, holding the fields vtable in our own structs is wrong (not
guaranteed to be safe by the plugin API), so don't do that.