new(plugin): introduce concept of associated data
Values held in tables can now have extra data associated with them.
The data is not passed to the API in any way and has to be fully
managed on the SDK side.

Currently, associated data isn't used yet (all relevant trait
implementations have `AssocData = ()`), but it's important for
supporting nested tables with the derive macro: we need to store
the field descriptors somewhere and AssocData is the place.