new(plugin): introduce metadata concept
For the upcoming derive macro, introduce a concept of metadata
for imported tables. `Table` values now have an associated metadata
type, that's shared between the table and all its entries (to be
precise, it's cloned, but we use Rc, so it's a refcount).

We need this to share the dynamic field descriptors across all
instances of each nested table.