update(plugin): make TableFields hold a lifetime
Even though the SDK hands out only references to a TableFields,
the struct itself is valid only for the duration of an API call,
so making it hold the lifetime of a reference to the actual vtable
makes it harder to misuse.

It also allows us to change the struct to hold a reference
to the vtable internally, which we'll do in a moment.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>