fix(plugin): make Vtable Send+Sync when needed
The Vtable struct only describes our own tables, which are safe
to use across threads with the thread-safe-tables feature.

We need Send+Sync on the Vtable struct to support nested tables
in a multithreaded scenario and it should be safe enough since
for a particular Table instance, the Vtable object never changes
and the function pointers inside point to thread-safe methods.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>