cleanup(event): remove the Borrowed trait
The purpose of this trait was to get from a borrowed type to
an owned type. The only use was to figure out the right type
for the owned variants' fields, but since we already have
a whole module with the owned types, just use that directly.

Cuts down build times a bit more still.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>